Fixed a bunch of Clang static analyzer issues.
As part of this, -[ShipEntity newMissile] was renamed to -selectMissile, because the "new" prefix conventionally indicates a method that returns a new object which is not autoreleased (as in -[Universe newShipWithRole:]).

Also introduced a bunch of annoying but technically correct "potential leak" issues. (These all refer to cases where an object is weakRetained and passed to JS_SetPrivate(). They aren't actual leaks, but they don't follow normal OpenStep retain/release behaviour.)"